pagination fixes
